image Provider improvement
