+++ src/components/ReactComponent/blog/enhancements/diagram-renderer/MermaidRenderer.tsx
@@ -1,4 +1,4 @@
-import React, { useEffect, useMemo } from "react";
+import { useEffect } from "react";
 import { memo } from "react";
 import { useMermaidRenderer } from "./hooks/useMermaidRenderer";
 import { useDiagramRenderer } from "./hooks/useDiagramRenderer";
