+++ src/content/blog/client-side-mermaid-diagrams-astro-react.mdx
@@ -155,6 +155,8 @@ Now, whenever a blog post has Mermaid code blocks, the following happens:
 5. React buttons (copy, fullscreen) are injected into each SVG container.
 6. The entire process runs _after_ the page has loaded, without blocking the initial render.
 
+---
+
 ## Final Thoughts: When to Stop Fighting the Build System
 
 This whole experience taught me a valuable lesson:
